Some minor fixes
Fix defaults for serverUrl and username so that we can properly detect
when they are set.

Allow stageDir to be augmented with a context directory in order to
support projects that want to build more than one image.

Fix some comments (Dockers use of registry and repository continues to
confound me).