Fixed baseImage overriding
Overriding the default base docker image now works as intended. It can
be overriden for all DockerTasks using the docker extension and it can
be overriden on a per task basis.

Also the logic to determine the default image based off the Java target
version has been extracted into a seperate JavaBaseImage enum.

Furthermore the convention task 'distDocker' is now treated the same as
any other Docker task in terms of default base image detection.

The added tests in the DockerTaskTest class hopefully prevent future
breakage of this logic.