Add support for running Docker containers
As part of this I created a common super-class used by the now 2 gradle
tasks.

Added tests (though testing of the run command is non-existant, still
figuring that one out)

Fixes to build files post rebase

Expand tests and fix a couple of issues found

WIP - fix relating to run with volumes

Move to released docker-java code and some formatting tweaks

Fixed 2 bugs found on Linux

More fixes to the native client

The image name (aka "tag") needs to be the last element of the command
line.

The separator for most of the key value arguments is a ':', though for env
we need an '='.

Complete previous fix.